Move implementation of ank set state into own file
Issue-Id: #229
